Start using Pydantic and directly talk to Canvas
I'm going to transition away from the canvasapi package. It's a great tool to get started with talking to Canvas, but I find myself clicking through to the official Canvas API docs all the time and when you want to include some additional paramters they are only returned as a dictionary and not objects. So you end up with a mixed object with attributes + dictionary thing which is... not great. I want to have more control and I want everything to be an object. So I'll start using httpx to talk to the API and Pydantic to turn the response into full Python objects.